Refactor header navigation into a separate component
