+++ web-app/src/app/(landing)/components/header.tsx
@@ -1,12 +1,9 @@
-"use client";
-
-import { ChevronDown } from "lucide-react";
 import Link from "next/link";
-import { usePathname } from "next/navigation";
 
 import { SokosumiLogo } from "@/components/masumi-logos";
 import { Button } from "@/components/ui/button";
-import { cn } from "@/lib/utils";
+
+import Navigation from "./navigation";
 
 export default function Header() {
   return (
@@ -34,71 +31,6 @@ export function AuthButtons() {
       <Link href="/signup">
         <Button>Sign Up</Button>
       </Link>
-    </div>
-  );
-}
-
-function Navigation({
-  className,
-  ...props
-}: React.HTMLAttributes<HTMLElement>) {
-  const pathname = usePathname();
-
-  return (
-    <div className="flex items-center gap-8">
-      <nav
-        className={cn("flex items-center space-x-4 lg:space-x-6", className)}
-        {...props}
-      >
-        <Link
-          href="/gallery"
-          className={cn(
-            "inline-flex items-center gap-1 text-sm font-medium transition-colors",
-            pathname === "/gallery"
-              ? "text-primary"
-              : "text-muted-foreground hover:text-primary",
-          )}
-        >
-          Agents Gallery
-          <ChevronDown className="h-4 w-4" />
-        </Link>
-        <Link
-          href="#how-it-works"
-          className={cn(
-            "inline-flex items-center gap-1 text-sm font-medium transition-colors",
-            pathname === "/#how-it-works"
-              ? "text-primary"
-              : "text-muted-foreground hover:text-primary",
-          )}
-        >
-          How it works
-          <ChevronDown className="h-4 w-4" />
-        </Link>
-        <Link
-          href="#join-our-community"
-          className={cn(
-            "inline-flex items-center gap-1 text-sm font-medium transition-colors",
-            pathname === "/#join-our-community"
-              ? "text-primary"
-              : "text-muted-foreground hover:text-primary",
-          )}
-        >
-          Community
-          <ChevronDown className="h-4 w-4" />
-        </Link>
-        <Link
-          href="#monetize"
-          className={cn(
-            "inline-flex items-center gap-1 text-sm font-medium transition-colors",
-            pathname === "/#monetize"
-              ? "text-primary"
-              : "text-muted-foreground hover:text-primary",
-          )}
-        >
-          Monetize
-          <ChevronDown className="h-4 w-4" />
-        </Link>
-      </nav>
     </div>
   );
 }
